Split out placement dates
In order to prevent dependency cycles
