+++ server/utils/match/index.test.ts
@@ -32,7 +32,6 @@ import {
   keyDetails,
   lengthOfStayRow,
   mapUiParamsForApi,
-  placementDates,
   placementLength,
   placementLengthRow,
   placementRequestSummaryListForMatching,
@@ -227,19 +226,6 @@ describe('matchUtils', () => {
   describe('placementLength', () => {
     it('formats the number of days as weeks', () => {
       expect(placementLength(16)).toEqual('2 weeks, 2 days')
-    })
-  })
-
-  describe('placementDates', () => {
-    it('returns formatted versions of the placement dates and durations', () => {
-      const startDate = '2022-01-01'
-      const lengthInDays = '4'
-
-      expect(placementDates(startDate, lengthInDays)).toEqual({
-        startDate: '2022-01-01',
-        endDate: '2022-01-05',
-        placementLength: 4,
-      })
     })
   })
